添加全局异常统一处理类，及处理mapper返回Map是字段为空的配置及配置类

+++ src/main/java/com/wzj/service/AbstractService.java
@@ -35,7 +35,7 @@ public abstract class AbstractService<T> {
         mapper.insertList(models);
     }
 
-    public void deleteById(Integer id) {
+    public void deleteById(Long id) {
         mapper.deleteByPrimaryKey(id);
     }
 
@@ -47,7 +47,7 @@ public abstract class AbstractService<T> {
         mapper.updateByPrimaryKeySelective(model);
     }
 
-    public T findById(Integer id) {
+    public T findById(Long id) {
         return mapper.selectByPrimaryKey(id);
     }
